Fixes an issue dropping a gotten preference value
The old code refered a loop counter 'key' in a fulfilled function
called by a Promise returned from get_pref().
'key' was undefined when the promise was fulfilled, then
the fulfilled function could not used the correct key value.

This fixes the issue by making a lamda function which uses
the correct key.

'pref_set' is also handled like 'key', but it is just cleanup and
the change does not impact to 'pref_set', since it is a pointer,
not immediate object.

This also moves the fulfilled function into get_pref().