core: Do not doubly call init() and fini()
Twice calling add_view() caused double initialization and finilization.
It caused "Index" theme shows the last column in the middle of dummy columns.

I did not like showing "Bookmark Bar" category as the first one in the case
that an user adds a bookmark in the folder "Bookmark Bar".

I changed my mind.

It is OK since I don't add a bookmark with such style.
Bookmarkviwer does not show category name if all children of a category
are folder. Then I am unnecessary to see the category "Bookmark bar".